[refactor]: profilebar hoursago int -> string으로 수정(#37)
