+++ app/src/main/java/com/texthip/thip/ui/mypage/component/SavedFeedCard.kt
@@ -126,7 +126,7 @@ private fun SavedFeedCardPrev() {
         userRole = stringResource(R.string.influencer),
         bookTitle = "책 제목",
         authName = "한강",
-        timeAgo = 3,
+        timeAgo = "3시간 전",
         content = "무한대로입력가능합니닷무한대로입력가능합니닷무한대로입력가능합니닷무한대로입력가능합니닷무한대로입력가능합니닷무한대로입력가능합니닷무한대로입력가능합니닷무한대로입력가능합니닷",
         likeCount = 10,
         commentCount = 5,
@@ -142,7 +142,7 @@ private fun SavedFeedCardPrev() {
         userRole = stringResource(R.string.influencer),
         bookTitle = "책 제목",
         authName = "한강",
-        timeAgo = 3,
+        timeAgo = "3시간 전",
         content = "한줄만 입력 가능",
         likeCount = 10,
         commentCount = 5,
