+++ app/src/main/java/com/texthip/thip/ui/signin/screen/SigninNicknameScreen.kt
@@ -1,4 +1,4 @@
-package com.texthip.thip.ui.signin
+package com.texthip.thip.ui.signin.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -15,7 +15,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.texthip.thip.R
 import com.texthip.thip.ui.common.forms.FormTextFieldDefault
-import com.texthip.thip.ui.common.topappbar.DefaultTopAppBar
 import com.texthip.thip.ui.common.topappbar.InputTopAppBar
 import com.texthip.thip.ui.theme.ThipTheme.colors
 import com.texthip.thip.ui.theme.ThipTheme.typography
